+++ src/commands/index.ts
@@ -1,0 +1,3 @@
+export * from './run';
+export * from './setup';
+export * from './update-integrations';